Add dynamic configuration docker compose with Jinja template

+++ script/clean.sh
@@ -7,6 +7,14 @@
 cd "$(dirname "$0")"
 cd ..
 
+read -p "Do you want to remove virtual python environment "\
+"(./env folder)? [Y/n]: " \
+answer
+if [ "${answer}" != "Y" ]; then
+  exit 0
+fi
+rm -v -rf ./env
+
 read -p "Do you want to remove services "\
 "(docker containers, images and network)? [Y/n]: " \
 answer
@@ -32,5 +40,13 @@ if [[ -z "${network_id}" ]]; then
 else
   docker network rm ${network_id}
 fi
+
+read -p "Do you want to remove generated configuration "\
+"(docker-compose.yml file)? [Y/n]: " \
+answer
+if [ "${answer}" != "Y" ]; then
+  exit 0
+fi
+rm -v docker-compose.yml
 
 echo "You must remove your minimal-<os> image at the hand, if you want..."
